feat: print time in logger (#36)

+++ kernel/time/time_manager.cpp
@@ -15,7 +15,7 @@ double influx::time::time_manager::seconds() const {
 double influx::time::time_manager::milliseconds() const {
     kassert(_timer_driver != nullptr);
 
-    return (double)_timer_driver->current_count() / (_timer_driver->count_frequency() / 1000.0);
+    return (double)_timer_driver->current_count() / ((double)_timer_driver->count_frequency() / 1000.0);
 }
 
 void influx::time::time_manager::tick() const {}
